Added ColorTile class
Updated all the Tile classes with a Grid world variable. This allows
Tiles to place colors in the Grid. In the ColorTile,everytime the color
updates it tries to copy itself in the direction its required to go. if
it cannot update it stops e.g. hits a wall or solid block.

Rawhide out