Added Metadata Processing to Solid, Receptor, Projector, Detector.
Added an abstract method to Tile -- public void decodeMetaData(short
metaData) -- and implemented it in to the classes mentioned. Updated the
Projector and Detector. Removed the side booleans and replaced them with
our direction enum. Added a color support for Detector. Currently
storing a short and not a java.Color class.

Rawhide